Resolves #2031 Replace Reselect with Memoize Remove package Reselect Delete file selectors/dependencies Add the memoize function to components/DependencyGraph/index.tsx
