+++ packages/jaeger-ui/package.json
@@ -93,7 +93,6 @@
     "redux-first-history": "^5.1.1",
     "redux-form": "^8.3.10",
     "redux-promise-middleware": "^6.1.3",
-    "reselect": "^4.1.6",
     "store": "^2.0.12",
     "ts-key-enum": "^2.0.0",
     "tween-functions": "^1.2.0",
