+++ packages/jaeger-ui/tsconfig.lint.json
@@ -49,7 +49,6 @@
     "src/reducers/index.js",
     "src/reducers/services.js",
     "src/reducers/trace.js",
-    "src/selectors/dependencies.js",
     "src/utils/configure-store.js",
     "src/utils/sort.js",
     "package.json"
